refactor: move internal packages to pkg directory
